Deleted the actual "cost function" code
Now that V1 doesn't use the cost function, but V2 does, V1 doesn't require it anymore and it was causing confusion. Back-propagation is done directly from aimed values - so making large decisive changes in weights (as we are dealing with aims of 0 or 1). Therefore quite efficient, however as mentioned before, not practical for real life cases. Also deleted variable call c, as it is no longer needed.